[FIX] Enclosures/Services lists
