+++ app/src/main/java/fr/isen/goofyzoo/screens/services/ServiceDetailScreen.kt
@@ -1,0 +1,43 @@
+package fr.isen.goofyzoo.screens.services
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+import fr.isen.goofyzoo.models.Service
+
+@Composable
+fun ServiceDetailScreen(navController: NavHostController) {
+    val service = navController.previousBackStackEntry?.savedStateHandle?.get<Service>("service")
+    service?.let {
+        ServiceItem(service)
+    }
+}
+
+@Composable
+fun ServiceItem(service: Service) {
+    Column(
+        modifier = Modifier
+            .padding(16.dp)
+            .background(MaterialTheme.colorScheme.surface, shape = RoundedCornerShape(8.dp))
+            .padding(16.dp)
+    ) {
+        Text(
+            text = service.name,
+            color = MaterialTheme.colorScheme.primary
+        )
+        Spacer(modifier = Modifier.height(8.dp))
+        Text(
+            text = "ID Service: ${service.id_service}",
+            color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.7f)
+        )
+    }
+}
